rewrite to use powershell
Original version of this module used pure `netsh` to enumerate and edit rules
which works great most of the time. Modern versions of Windows have rules that
are not resolvable or deletable using the `netsh` command (see #3).

To properly manage these rules its necessary to use PowerShell and this
required a significant rewrite. The `ps-bridge.ps1` script is used to manage
all interactions between Puppet and Windows in terms of enumerating, creating
and deleting rules. This allows the script to be edited and tested separately
and prevents vendor lock-in.

The module is now enabled for acceptance testing using PDQTest 2 and docs
have been updated.

Note that puppet type interfaces have changed and end-user code will require
updating to work with the new module.