Handle rule names that do not resolve
Issue #10 was caused when there are rules present which have no resolvable
name in the windows registry. This may or may not indicate registry corruption.

To handle this, just use the raw name, eg

```
@{Microsoft.Windows.Apprep.ChxApp_1000.14393.2580.0_neutral_neutral_cw5n1h2txyewy?ms-resource://Microsoft.Windows.Apprep.ChxApp/resources/DisplayName}
```

As the display name in puppet. This is the same behaviour as Windows Advanced
Firewall GUI